Now reads decoded data when `DecryptMessage` wants to renegotiate
It appears that `DecryptMessage` can still result in decoded data when it
signals that a renegotiation needs to occur. Originally, these bytes
were being discarded, which could cause the connection to block
indefinitely on a read. This only seemed to occur when a client was
using a certificate to connect to a server.